GUI_3
This version has an expanded capability of what I put up last night. The
search button now returns the values that the user entered into text
boxes, and saves them even after the window has been closed. I'm still
working on adding a button which adds an extra text field; the button is
there but the call back isn't done.

To use these functions, you need an an instance of the BoxCreation
class, and call the functions on that. I think for pretty much all of
the functions you also need to supply generic labels, buttons, etc, as
parameters, like b = Button().  I'll comment it up tomorrow, but
hopefully the names of the parameters make it somewhat clear what is
needed for each function.