Hopefully useable if you guys require these classes/functions
I think that this does all of the basic functions, albeit the
presentation of the box is still disorganized; the widgets are put on
the box in the order in which their functions were called. I'm working
on the URL check as well as the error boxes which get raised, but
hopefully this will give you guys enough if you've been waiting for this
stuff. IMPORTANT: comment out the import for httplib2 at the top of the
valueHolder file, if you haven't downloaded it already. I think I did
this for the first class, but not valueHolder. I'll be working on this
for the next several hours, so email me if you find something which
doesn't work right. I'll upload what will be the final version sometime
tonight.